long to int

+++ help-desk-api/Models/Dto/Menus/UserMenuDto.cs
@@ -36,7 +36,7 @@ namespace Models.Dto.Menus
         public string UserName { get; set; }
         public string ImagePath { get; set; }
         public int RoleId { get; set; }
-        public long CompanyId { get; set; }
+        public int CompanyId { get; set; }
  
     }
     public class UserBiReportDto
@@ -71,9 +71,9 @@ namespace Models.Dto.Menus
 
     public class PermittedActionsOutputDto
     {
-        public long UserId { get; set; }
+        public int UserId { get; set; }
         public int MenuId { get; set; }
-        public long MapId { get; set; }
+        public int MapId { get; set; }
         public string MenuName { get; set; }
         public string SystemMenuName { get; set; }
         public string HttpVerb { get; set; }
