+++ help-desk-api/Models/Entities/Notification/NotificationTemplateModel.cs
@@ -4,12 +4,12 @@ using System.ComponentModel.DataAnnotations.Schema;
 namespace Models.Entities.Notification
 {
     [Table("NotificationTemplate", Schema = "notification")]
-    public class NotificationTemplateModel : BaseEntity<long>
+    public class NotificationTemplateModel : BaseEntity<int>
     {
-        public long TicketTypeId { get; set; }
+        public int TicketTypeId { get; set; }
         public NotificationTrigger Trigger { get; set; }
         public NotificationType NotificationType { get; set; }
-        public long? EmailConfigurationId { get; set; }
+        public int? EmailConfigurationId { get; set; }
         public string SubjectTemplate { get; set; }
         public string BodyTemplate { get; set; }
         public string CcList { get; set; }
